backend/requirements.txt
fastapi==0.104.1
python-dotenv==1.0.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
supabase==2.0.0
firebase-admin==6.2.0
google-cloud-documentai==2.20.1
google-cloud-aiplatform==1.70.0
vertexai==1.70.0
python-multipart==0.0.6